+++ NEWS.md
@@ -11,6 +11,7 @@
 ### Minor updates
 
 * Allow for checks for links.
+* Changed functions to allow for nodes without OPEX that are not `Availability` nodes and nodes without capacity that are not `Availability` nodes.
 
 ## Version 0.8.3 (2024-11-29)